Clarify argument usage.
Even when using my own script I couldn't help notice that the
--partitions cmdline argument was a bit misleading. It would
appear to indicate the number of partitions the user wants to
create, which isn't correct. So I changed it to --partpercents
because this argument allows the user to specify the number of
partitions by giving a list of their sizes in percentages.