Add ARM support.
I have been working specifically on creating images which will run
under Linaro's QEMU which has support for Gumstix's Overo board.

To boot with this virtual machine, the user needs to setup a 1st
partition as follows:
	- formatted as VFAT
	- partition ID set to 0x0c
	- contains the following data
		- MLO
		- u-boot
		- uImage

I don't know if these steps are Overo-specific or are general for
all ARM devices.